requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.40.0
google-api-python-client>=2.120.0
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
